homeauto/requirements.txt
async-timeout==4.0.2
click==8.0.4
Deprecated==1.2.13
Flask==2.0.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.7
python-decouple==3.6
redis==4.2.0
typing-extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.14.0
